Extract deletion to DeleteButton component
